Check dotnet without which or where first
Which and where can be slow to find. For the common case which we want to make fast, we can just check 'dotnet' first and see if that works.

Note: Ideally, this function would only return a string, but that would require writing a wrapper around this return type just for this function and does not follow the other existing patterns.

With respect to tests: This should be covered by the existing 'find path' tests.