Skip CHCP Calls and Utilize DOTNET_CLI_UI_LANGUAGE (#2233)
* Set DOTNET_CLI_UI_LANGUAGE instead of chcp

Calling chcp has significant performance impacts.

One problem is if the dotnet host does not respect the language setting and a user only has the host or runtime installed. We cannot perform the architecture check in this case.

Also adds NOLOGO to prevent the issue of first run exp + performance delay from it.

* Consider when the output architecture is empty