Handle Corrupt Install Script Installs (#2235)
* Check if the dotnet install is valid

The dotnet install script has a bug where it wont do the install if the folder is not empty. Sometimes it will leave a partially done install, like if killed, in which case we get to a bad state.

This fixes that by cleaning up the directory, and or checking if dotnet.exe exists, whether it works.
dotnet --list-runtimes is a fast way to do this since it should only invoke the host and not need to invoke the SDK. doing just 'dotnet' will not work as that does not return 0.

Also, gets rid of the 'lock' exception since thats from the proper-lockfile lib which we don't use anymore.

Resolves: https://github.com/dotnet/vscode-dotnet-runtime/issues/2059
First scrapped attempt: https://github.com/dotnet/vscode-dotnet-runtime/pull/2106

* Add missing await

* Recursively delete the stuff

wipe dir only wipes the files, and the script looks at the folders

* Remove tests for deprecated code

The Install Multiple versions tests fails bc the deprecated sdk code installs everything into the same folder, so this new cleanup logic will wipe the older folder.

The uninstall logic fails because the original sdk code does not work with respect to finding a local sdk when a global sdk is on the machine (it finds a 7.0.4xx sdk), this code is not used by anything anymore so it is likely not worth investigating.

* Update changelog

* Changelog should mention min vscode version

* Check if dotnet meets the requirement of the install