Handle Existing Installs with Other Request Orderings Correctly
A problem C# has noticed is that there is a small chance for the install script to think it is done installing, and for our extension to think we have installed, even though not all files are there.

Upon closer inspection, I realized some issues and resolved them:

# There is a race condition

We have a check to see if there is an existing 'promise' that is working on doing the same install as another request. However, sometimes that check will run simultaneously. What we did before was only lock during the process of running the install script, locking on the installation id (version+arch+install type+mode). Now we lock during the process of looking for existing installs, doing the install, and recording the install.

2. The install script handler used the wrong type (InstallationContext) instead of a AcquisitionContext to validate it, so the mode it was validating was undefined. This means it would consider a valid install to be invalid.