Dont use entire env as cache key (#2251)
The cache is a local memory cache reset upon vscode loading.
It caches commands that we run to avoid running them again.
We noticed the environment changes very often but most commands are only impacted by certain environment variables. If there is an env var we did not consider that could change a command's output, then this would make that case incorrect. However, we think the perf gain for this and the lack of opportunity for this is OK. For example, we check PATH, DOTNET_ROOT, etc and would still respond to those changes when we map a command to run to a cache key.

Possible commands are impacted, are dotnet --info, --list-runtimes, and querying the registry, as well as running netstat, checking etc/os-release, etc. I think the reg is the most concerning one as the PATH in vscode may not update there but it is on a 9 second timer and a correct env var already should not exist to make that update properly with the cache.