Remove promise and 'installing' tracking mechanism, prefer mutex
- Remove 'installing' state.

- Remove promises store.

The promises was a way to check if an existing function was running and executing the same install. This was added at the very beginning. However, it was stored in local memory. This is not good as it is not persistent across multiple instances of vscode running at the same time. Now we have a mutex over the local storage folder that works cross-process and intra-process. The mutex should protect the entire install operation. Other similar installs will wait for the mutex to be free (or if the application dies.) Then if it is valid they will see its an existing install and validate it, then use it. If not, the invoker now wipes the install folder.

Now, one problem is validation only checks --info arch and --list-runtimes or sdks and that it matches the version requirement, as well as that the dotnet host files and main expected files are there. There is no easy way to fully validate that everything worked without enumerating every file and keeping a list of every valid file for every valid build, or rezipping it exactly how our tooling does and comparing the sha (possibly easier but still hard.) This is not a new bug that has been introduced, however, considering before the behavior must have also existed, because the 'installed' state key check only validated that the executable was present, and the check for partial installs only removed installs from the 'installing' list, not the 'installed' list.

Now, the 'installing' list is no longer relevant as it was only used for this promise holding based approach.