Uninstall Logic Handles Corrupt Save Data (#2166)
* fix typo and collect whether deletion worked

* Fix some of the deletion code

* Add event for when uninstall is external

* Handle Existing Installs with Other Request Orderings Correctly

A problem C# has noticed is that there is a small chance for the install script to think it is done installing, and for our extension to think we have installed, even though not all files are there.

Upon closer inspection, I realized some issues and resolved them:

# There is a race condition

We have a check to see if there is an existing 'promise' that is working on doing the same install as another request. However, sometimes that check will run simultaneously. What we did before was only lock during the process of running the install script, locking on the installation id (version+arch+install type+mode). Now we lock during the process of looking for existing installs, doing the install, and recording the install.

2. The install script handler used the wrong type (InstallationContext) instead of a AcquisitionContext to validate it, so the mode it was validating was undefined. This means it would consider a valid install to be invalid.

* Fix test logic

removes tests for deprecated education bundle extension

* Fix multiple versions  test

* Fix null arch

* Code cleanup

- Movev the lock for partial into the local since it already needed to do so, and this would be a perf hit to acquire it twice

* Remove promise and 'installing' tracking mechanism, prefer mutex

- Remove 'installing' state.

- Remove promises store.

The promises was a way to check if an existing function was running and executing the same install. This was added at the very beginning. However, it was stored in local memory. This is not good as it is not persistent across multiple instances of vscode running at the same time. Now we have a mutex over the local storage folder that works cross-process and intra-process. The mutex should protect the entire install operation. Other similar installs will wait for the mutex to be free (or if the application dies.) Then if it is valid they will see its an existing install and validate it, then use it. If not, the invoker now wipes the install folder.

Now, one problem is validation only checks --info arch and --list-runtimes or sdks and that it matches the version requirement, as well as that the dotnet host files and main expected files are there. There is no easy way to fully validate that everything worked without enumerating every file and keeping a list of every valid file for every valid build, or rezipping it exactly how our tooling does and comparing the sha (possibly easier but still hard.) This is not a new bug that has been introduced, however, considering before the behavior must have also existed, because the 'installed' state key check only validated that the executable was present, and the check for partial installs only removed installs from the 'installing' list, not the 'installed' list.

Now, the 'installing' list is no longer relevant as it was only used for this promise holding based approach.

* Add public command to reset data and uninstall all

Also makes uninstall all hold the lock for 'installed' so it cannot cause a weird partial install state

* Consider test does not install

* Fix Lint

* Fix extra !

* Post status resolved

it seems to resolve the path and get the existing install but then it fails anyways somehow. assertAcquisitionSucceeded may be failing.

* Add test for 'resetData' command

* Fix return type

I forgot to save :(

* Code style improvement

* fix test timeout

* remove extra semicolon xd

* Add file is busy check

* Add in use check

* fix import

* Remove unnecessary information from log

* Test should use finally

* Check each folder for a dotnet executable

readdir gets the directories and files. fixes impl of wipe dir -- it should not have any side effect in other cases ( tested )

* Always close file handle

* use correct file handle type

* File is open check should hold exclusive fd

* Invoker also checks if in use before wiping

* update test msg

* promisify open check

* silly promisify things

* fix catch

* missing }

* lint

:yawn:

* Add test for uninstall and uninstall All

* Fix reset data command

* Fix test and dont print output twice

* Return completed event to stop output on success