Quick Attempt To Preemptively Load Cache
Calling dotnet --info, --list-runtimes etc is time consuming. We can do this at activation instead of later making assumptions about what most callers will request to speed this up. This does cause some extra work to be done, so this is an experiment.

We can more intelligently find the common case in the future.