Skip checking install requests for known extensions
We check the package.json file for each extension which is an expensive wholistic file read that happens async during startup. This is slow and we can skip that for many extensions that we know will never need to ask us to install or find dotnet.

What I did to generate this list:
- Look at the extensions.all output on an empty (no custom extensions) code
- Look at the marketplace and add all first party / business partner related extensions I saw on the page that had the highest number of downloads

This is not an endorsement of these extensions or their codebases.
There is not a way to filter these or get anything besides .all of the extensions at this time.