Fix logic for rollFoward options
It assumed that latestPatch and latestFeature could not be fulfilled if there was no patch requested by the user. But we should instead accept it. This would make '7.0.20' get rejected over '7.0' with 'latestPatch' which doesnt make sense because it should accept any later patch over the specified patch. In this case it was not specified so it should be accepted.

Logic extracted out into another function was not changed.