Don't find Locally Managed Runtimes (#2302)
* Don't find Locally Managed Runtimes

If we change this, everyone who calls the findPath API will stop getting runtime updates, because they call `findPath` before `acquire`. We need to update the runtimes we install (the users own managed runtimes are different) so we can't do this yet. Let's take this out but leave in the feature so we can just revert the lookup and then revert this part later after we implement the update story.

That will cause a big breaking change because it will change the surface area of the acquire API to no longer update by default but instead update over time periodically.

* Update vscode-dotnet-runtime-extension/src/test/functional/DotnetCoreAcquisitionExtension.test.ts