Update Dependencies, Migrate to Node 22
VS Code 1.100.0 migrated from node 20 to node 22. We should stay aligned with VS Code. Because our binaries actually run under whichever version node in Code is. This can cause extreme issues such as a complete failure to load if the compiler or language feature changes, and we would not know for some time.