Find Locally Managed Runtimes
This was implemented but then backed out for release in https://github.com/dotnet/vscode-dotnet-runtime/pull/2302.

This includes one fix which is to respect DOTNET_ROOT over the locally installed runtimes. Users who set DOTNET_ROOT typically have an explicit scenario that they want.

This was backed out because we need to do auto update of the local installed runtimes for this feature to be compatible with the API shape of keeping customers secure. We will do that in this release.