Fix the error output
The true problem was that the error type was bundled inside another error type due to some old architectural decision. CoPilot had put a band aid over this issue and it was only partially fixed. I took the liberty of improving the error messages as well. We shouldnt include a stack in a publicly facing message.