Add fallback for /usr/lib/os-release to detect distro
More context at https://github.com/dotnet/vscode-dotnet-runtime/issues/2337, but TLDR is in the code comments which is that this is recommended as an official fallback. Note that there are still distros which do not adhere to this specification but the ones we support 1st party do, and if the distro cannot be detected, the unsupported message is still given.

This also changes the detection, I believe it was written using 'cat' because a different approach from etc/os-release was first attempted to try to work across further platforms. Anyway, its very expensive to run cat for this so let's not do that.

There may be a string parsing error, I will see if the tests pass before I spend the time to go try this on a VM.