Add setting to disable messages in the output window (#2332)
* Add setting to disable messages in the output window

Resolves https://github.com/dotnet/vscode-dotnet-runtime/issues/2027

The output window in vscode, under the '.NET Install Tool' tab may be annoying/unnecessary to the user. This adds a setting to disable that output.

* Interface as OutputChannel

VS Codes impl may not return the thread which would make the inner test loop of mocha never complete so the test hangs indefinitely

* Dont use acquisition started

We can check started and completed once, no need to do this, its stopping the event loop from finishing

* Reduce setInterval Speed

We dont need to print out that often and take thread resources like that.

* Add the setting

* Fix the setting

* fix ,

...