Add `dotnetSDKPath` Setting for User SDKs and Projects
The .NET `dotnet.findPath` API provided by the .NET Install Tool only uses machine wide SDKs at this time.

This API is used by C# and C#DK, as well as others, to resolve the .NET SDK to use for the project system. We can update this to better enable local SDK or custom SDK scenarios. In general, the PATH can be updated to an admin SDK, but users may get confused if they have multiple SDKs on the PATH, e.g. with multiple lines in their rc profile, which has happened internally. In addition, DOTNET_ROOT and other variables impact the selection of DOTNET, and this may further complicate the issue, but DOTNET_ROOT is only applicable to the Runtime selection.

In the past, `dotnetPath` was supported by C#/OmniSharp but it was deprecated.
I'm unsure of the name 'dotnetSDKPath' : unfortunately we can't change the old setting, `existingDotnetPath`.

Caveats:
`existingDotnetPath`: Is a path to the dotnet.exe that should have runtime beside it, and is used to run the code that is shipped within C#/C#DK and other extensions.
`dotnetSDKPath`: Is a path to the dotnet.exe that should have an SDK beside it (and therefore also an SDK), that should be used by the project system and CPS to run the users project. We may need to have further work from the debugger team or coordinate with them as they may have their own setting.

This does allow a user to override the system PATH and use a local SDK which may not be secure or up to date in an enterprise scenario. However, VS Code generally does not run in an elevated context and suggests against doing so. We should ensure that elevation is not leveraged by the internal calling APIs, and I dont believe that is the case.