acquireStatus aware of offline Installs
The `acquireStatus` API does not use the offline detection logic to fallback to installed runtimes. This API in general is not needed anymore but it looks like we still have some API dependents and it was never officially deprecated, so we should support that.

This reimplements the same logic to do so and should help resolve https://github.com/dotnet/vscode-dotnet-runtime/issues/2325