WIP - Add test via CGP Sonnet 4
Can we add a test that checks that the worker can find existing installs if the user is offline? To do this, use a Mock web worker which implements isOnline and says false always. The install tracker will also need to be mocked. You may need to change the interface of the functions to take in (or create by default to the existing behavior) the web worker or install tracker singleton dependency.

This is getSimilarExistingInstall.

Files: MockObjects
DotnetCoreA..Worker
WebR....Ton.ts
InstallT...Ton.ts
CoreWorke....Test.ts