Create API to get output of --list-runtimes and --list-sdks with architecture
The motivation to do this is described in https://github.com/dotnet/vscode-dotnet-runtime/issues/2315 as well as the spec.

This will enable the caching and performance related code to be shared and was mainly implemented for the debugger so they can remove their call to dotnet --info.

This commit is step 1, some refactoring would be ideal.