Migrate to Deployment Model
The timed task will cause the pipeline to fail. It may have a 'resume' feature we can check to see if it actually failed or not, but there is no way to pause the next task and 'pass' that 'resume' onto the next step, because the timed task is using a 'serverless' pool.

This method is more complicated but a better e2e approach.
This commit removes the line which does the publish so I can test that the deployment 'gate' in the environment actually works.