Track installs returned this session
If we have returned a dotnet install / exe to use for that session of vscode we certainly dont want to uninstall it because its almost certainly in use, or will be used again

This is a somewhat poor design pattern and tight coupling. It sucks that any time we return an install we need to make sure we call this function. Maybe we could create a wrapper to return a dotnetPath object that registers this. Simultaneously though, folks would still need to know to call the wrapper.