Install Records should only be cleared after uninstall success
+ we should track an install if it was existing and we now have an owning extension upon it via acquire. We could do the same for findPath but we dont manage those installs and dont want to muddle the data with installs we own and ones we dont own.

There is a large flaw / added layer of complexity having 2 systems here.

With the sessions system, we can't track whether an install is no longer in use by multiple dependent extensions, only if the single one that's asking to uninstall is no longer dependent.

We could try to merge the system s.t. with each record we keep a list of possibly live session ids. But then we'd need to include the path in the record data. We should have done this from the beginning. But with older ext versions, we dont have a complete enough picture to correctly generate the path in all cases, so this would break upgrading from an old vers to the new version.