Add ability to acquire specific versions
We need to acquire older versions to test the functionality of update.
This also makes force be used so we always acquire in the sample extension, so vendor testing works as expected. We provide a new command for vendors to use to test manually.

We also add logging details so it's easier to diagnose issues. I've noted that we might be uninstalling despite the sample being a live dependent.