Force Uninstall should not be used, and we should clear old sessions
The logic to speed up search doesn't matter much since there will likely ever only be a few sessions open at a time (you need multiple vscodes/insiders/ code forks to run per session) and we dont want to bloat with a bunch of old sessions that never get removed, so fixed this.

We also don't want to force remove during auto update, so fixed this.

uninstall should not be an error if it's in use, so fixed this as well.

sanitization helpers fix should not impact this but ignoring non truthy values could cause issues if it's used elsewhere