Don't check host architecture if no installs are available
When users dont have dotnet installed, this resulted in an error message saying we couldn't resolve the host architecture. The host architecture failing to resolve is unexpected, but if there is no host then it's not unexpected. We shouldn't even take the cost to do the check when .NET is not installed.