Add support for dotnet 10 and newer distro versions
This is our version support system which powers our warnings for when users try to install out of support versions or not yet released versions for linux distros.

Ideally, this would be automated based on the documentation, though this likely won't be bumped in priority.

To try to minimize impact in the future, I've added the future versions of distros.

Resolves https://github.com/dotnet/vscode-dotnet-runtime/issues/2519